Move _set_data method to MarkerView.
